`@ssc/core`: organize commands
